+++ frontend/src/pages/user/settings/SettingsOther.jsx
@@ -2,18 +2,26 @@ import { useNavigate, useOutletContext } from 'react-router-dom';
 import usersService from '../../../services/users';
 import { useDispatch } from 'react-redux';
 import { resetAll } from '../../../reducers/auth';
+import { useState } from 'react';
 
 const SettingsOther = () => {
   const { user } = useOutletContext();
   const dispatch = useDispatch();
   const navigate = useNavigate();
 
+  const [loading, setLoading] = useState(false);
+
   const handleDelete = async () => {
+    if (loading) return null;
+    setLoading(true);
+
     if (confirm('Are you sure you want to delete your user?')) {
       await usersService.deleteUser(user.username);
       dispatch(resetAll());
       navigate('/login', { replace: true } );
     }
+    
+    setLoading(false);
   };
 
   return (
